neuron-framework-t1/requirements.txt
weave-python>=0.50.0
openai>=1.0.0
anthropic>=0.25.0
asyncio
aiohttp
pillow
numpy
pytest
pytest-asyncio
python-dotenv
pydantic>=2.0.0
typing-extensions